requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jose==1.0.0
MarkupSafe==3.0.3
passlib==1.7.4
pg8000==1.31.5
pillow==11.3.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
python-vlc==3.0.21203
rsa==4.9.1
scramp==1.4.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0
